refactor(spec): split the 800-line pty.go by concern (exec, mouse, send, keys)
