+++ internal/spec/pty.go
@@ -1,17 +1,5 @@
 package spec
 
-import (
-	"bytes"
-	"fmt"
-	"math"
-	"slices"
-	"strings"
-	"time"
-
-	"github.com/goccy/go-yaml"
-	"github.com/goccy/go-yaml/ast"
-)
-
 // ClearEnvEnabled reports whether the pty step opts into a cleared environment (#16).
 func (p *PTY) ClearEnvEnabled() bool { return p.ClearEnv != nil && *p.ClearEnv }
 
@@ -89,107 +77,6 @@ type PTYAction struct {
 	ExpectScreen *PTYExpectScreen `yaml:"expect_screen,omitempty"`
 }
 
-// PTYExec is one host command run mid-session (#380). It accepts either a
-// scalar string (argv-parsed like run.command) or a mapping carrying the shell
-// and timeout knobs.
-//
-// The command runs in the scenario workdir with the same environment the pty
-// child received, so `sandbox_home` / `clear_env` isolation is not quietly
-// punctured by the helper. Its output never joins the transcript — the
-// transcript is what the TERMINAL showed — and is kept only to explain a
-// failure. A non-zero exit, a timeout, or a failure to start is a hard step
-// error: the command is scaffolding, not the subject under test, so a broken
-// one must stop the run rather than leave the following expect_screen waiting
-// for a change that was never made.
-//
-// It is deliberately not a scripting language: a fixed command at a fixed point
-// in the session, no branching, no captured output feeding later steps.
-type PTYExec struct {
-	// Command is the program to run. Required.
-	Command string `yaml:"command"`
-	// Shell runs Command through the shell, like run.shell.
-	Shell *bool `yaml:"shell,omitempty"`
-	// Timeout bounds this command as a Go duration (default 10s). The remaining
-	// session budget bounds it too, whichever is shorter.
-	Timeout string `yaml:"timeout,omitempty"`
-
-	// mapped records that the author used the mapping form, so an empty mapping
-	// ({}) is rejected rather than read as "no command".
-	mapped bool
-}
-
-// ShellEnabled reports whether the exec command runs through the shell.
-func (e *PTYExec) ShellEnabled() bool { return e.Shell != nil && *e.Shell }
-
-// DefaultPTYExecTimeout bounds a mid-session host command when the spec does
-// not (#380). Scaffolding a session waits on should be quick; a command that is
-// not is one the author should have to bound on purpose.
-const DefaultPTYExecTimeout = 10 * time.Second
-
-// UnmarshalYAML decodes exec as a scalar command string or a mapping, rejecting
-// unknown mapping keys (a custom unmarshaler bypasses the loader's strict
-// decode). It decodes from the AST node so every shape error carries the
-// offending value's [line:col].
-func (e *PTYExec) UnmarshalYAML(node ast.Node) error {
-	fail := func(format string, args ...any) error {
-		return &yaml.SyntaxError{Message: fmt.Sprintf(format, args...), Token: node.GetToken()}
-	}
-	var one string
-	if err := yaml.NodeToValue(node, &one); err == nil {
-		e.Command = one
-		return nil
-	}
-	var raw map[string]any
-	if err := yaml.NodeToValue(node, &raw); err != nil {
-		return fail("exec must be a string or {command: ..., shell: bool, timeout: duration}")
-	}
-	e.mapped = true
-	for k, v := range raw {
-		switch k {
-		case "command":
-			str, ok := v.(string)
-			if !ok {
-				return fail("exec.command must be a string")
-			}
-			e.Command = str
-		case "shell":
-			b, ok := v.(bool)
-			if !ok {
-				return fail("exec.shell must be true or false")
-			}
-			e.Shell = &b
-		case "timeout":
-			str, ok := v.(string)
-			if !ok {
-				return fail("exec.timeout must be a duration string (e.g. \"10s\")")
-			}
-			e.Timeout = str
-		default:
-			return fail("exec: unknown key %q (accepted: command, shell, timeout)", k)
-		}
-	}
-	return nil
-}
-
-// MarshalYAML emits the shape UnmarshalYAML accepts, so a loaded exec
-// round-trips: the scalar form when nothing else was set, the mapping otherwise.
-func (e PTYExec) MarshalYAML() (any, error) {
-	if !e.mapped && e.Shell == nil && e.Timeout == "" {
-		return e.Command, nil
-	}
-	m := map[string]any{"command": e.Command}
-	if e.Shell != nil {
-		m["shell"] = *e.Shell
-	}
-	if e.Timeout != "" {
-		m["timeout"] = e.Timeout
-	}
-	return m, nil
-}
-
-// Label renders the exec symbolically for explain/doc (#380).
-func (e *PTYExec) Label() string { return fmt.Sprintf("exec %q", e.Command) }
-
 // PTYResize is a mid-session terminal resize (#379). Both dimensions are
 // required — there is no "keep the other one", because a spec that says what
 // the window becomes reads better than one that says what it changes by.
@@ -208,180 +95,6 @@ type PTYResize struct {
 	Cols int `yaml:"cols"`
 }
 
-// PTYSend is the polymorphic pty send payload (#26): exactly one of Text
-// (scalar form) or Key (mapping form) is set.
-type PTYSend struct {
-	// Text is sent verbatim; the empty string transmits EOF (^D).
-	Text *string
-	// Key is a named key, normalized to lower case.
-	Key string
-	// Times repeats a named key (#377), so sixteen presses of an arrow are one
-	// readable action instead of sixteen identical session entries. Zero and one
-	// both mean a single press; the repeats go out as ONE terminal write, which
-	// is what a held key looks like to the program reading it. Only valid with
-	// Key: repeating text is already expressible inline.
-	Times int
-	// Paste is text delivered as a BRACKETED PASTE (#378): the bytes go out
-	// wrapped in the markers a terminal puts around pasted input, so a program
-	// that distinguishes a paste from typing — a REPL that must not execute a
-	// pasted block line by line, an editor that must not auto-indent it — takes
-	// its paste path. The empty string is an empty paste, not the scalar form's
-	// EOF. Mutually exclusive with Key and with the scalar Text form.
-	Paste *string
-	// Mouse is a mouse event delivered as an SGR (1006) report (#381), for the
-	// TUIs that accept clicks and scrolling — lazygit, yazi, htop, fzf --mouse,
-	// anything on bubbletea. It is also sometimes the only sane way to reach a
-	// target: clicking a pane beats walking to it with twenty keystrokes.
-	// Mutually exclusive with Key, Paste, and the scalar Text form.
-	Mouse *PTYMouse `yaml:"mouse,omitempty"`
-}
-
-// PTYMouse is one mouse event (#381), encoded in xterm's SGR (1006) form.
-type PTYMouse struct {
-	// Row and Col are the 1-based screen cell, as a person would count them.
-	Row int `yaml:"row"`
-	Col int `yaml:"col"`
-	// Button is left (default), middle, right, wheel-up, or wheel-down.
-	Button string `yaml:"button,omitempty"`
-	// Action is click (default: press and release in one write), press, or
-	// release. A wheel button has no release, so it rejects one.
-	Action string `yaml:"action,omitempty"`
-	// Mods are shift, alt, and/or ctrl.
-	Mods []string `yaml:"mods,omitempty"`
-}
-
-// Mouse button and action vocabularies, and the SGR base codes each button
-// carries. Wheel buttons live at 64+ in xterm's encoding, which is why they are
-// not simply the next numbers after right.
-var (
-	ptyMouseButtons = map[string]int{
-		"left":       0,
-		"middle":     1,
-		"right":      2,
-		"wheel-up":   64,
-		"wheel-down": 65,
-	}
-	ptyMouseActions = map[string]bool{"click": true, "press": true, "release": true}
-	// Modifier bits, as xterm adds them to the button code.
-	ptyMouseMods = map[string]int{"shift": 4, "alt": 8, "ctrl": 16}
-)
-
-// PTYMouseButtonNames and friends spell the vocabularies for error messages.
-func PTYMouseButtonNames() string { return "left, middle, right, wheel-up, wheel-down" }
-func PTYMouseActionNames() string { return "click, press, release" }
-func PTYMouseModNames() string    { return "shift, alt, ctrl" }
-
-// ValidPTYMouseButton, ValidPTYMouseAction, and ValidPTYMouseMod report
-// membership in each vocabulary, so the loader rejects a typo instead of
-// sending an event the program cannot read.
-func ValidPTYMouseButton(name string) bool { _, ok := ptyMouseButtons[name]; return ok }
-func ValidPTYMouseAction(name string) bool { return ptyMouseActions[name] }
-func ValidPTYMouseMod(name string) bool    { _, ok := ptyMouseMods[name]; return ok }
-
-// IsWheel reports whether the button is a scroll wheel, which has no release
-// event in the SGR encoding.
-func (m *PTYMouse) IsWheel() bool {
-	return m.Button == "wheel-up" || m.Button == "wheel-down"
-}
-
-// ButtonOrDefault and ActionOrDefault fill in the documented defaults, so the
-// encoder and the validator agree about what an omitted field means.
-func (m *PTYMouse) ButtonOrDefault() string {
-	if m.Button == "" {
-		return "left"
-	}
-	return m.Button
-}
-
-func (m *PTYMouse) ActionOrDefault() string {
-	if m.Action == "" {
-		return "click"
-	}
-	return m.Action
-}
-
-// Bytes encodes the event as xterm SGR (1006) mouse reports: press is
-// "CSI < Cb ; COL ; ROW M" and release the same with a lowercase "m", where Cb
-// is the button's base code plus the modifier bits. A click is a press
-// immediately followed by its release, in one write, because that is what a
-// real click delivers and splitting it would let a redraw land between the two.
-func (m *PTYMouse) Bytes() []byte {
-	cb := ptyMouseButtons[m.ButtonOrDefault()]
-	for _, mod := range m.Mods {
-		// OR, not add: the modifiers are bit flags, so repeating one must not
-		// carry into the next bit. `mods: [ctrl, ctrl]` summed to 32, which is
-		// the MOTION bit — a report atago does not send and a TUI would read as
-		// something else entirely.
-		cb |= ptyMouseMods[mod]
-	}
-	press := fmt.Sprintf("\x1b[<%d;%d;%dM", cb, m.Col, m.Row)
-	release := fmt.Sprintf("\x1b[<%d;%d;%dm", cb, m.Col, m.Row)
-	switch {
-	case m.ActionOrDefault() == "press":
-		return []byte(press)
-	case m.ActionOrDefault() == "release":
-		return []byte(release)
-	case m.IsWheel():
-		// A wheel notch is a single event: there is no release to pair with it,
-		// and sending one would put a report on the wire that no real terminal
-		// produces. `click` on a wheel therefore means one notch.
-		return []byte(press)
-	default:
-		return []byte(press + release)
-	}
-}
-
-// DecodePTYMouseButton reverses the SGR button code back into the button name
-// and modifier names atago writes (#381), so `atago record --pty` can render a
-// captured report as the event it describes. It reports false when the code
-// carries a bit atago has no name for — motion reporting, a button beyond the
-// three it names — because a recording must replay the bytes it captured, and
-// naming only part of them would not.
-func DecodePTYMouseButton(cb int) (button string, mods []string, ok bool) {
-	// Peel the modifier bits off first; what remains has to be a button.
-	remaining := cb
-	for _, name := range []string{"ctrl", "alt", "shift"} { // high bit first
-		if bit := ptyMouseMods[name]; remaining&bit != 0 {
-			remaining -= bit
-			mods = append(mods, name)
-		}
-	}
-	// Report in the order the vocabulary documents, not the order peeled.
-	slices.Reverse(mods)
-	for name, base := range ptyMouseButtons {
-		if base == remaining {
-			return name, mods, true
-		}
-	}
-	return "", nil, false
-}
-
-// Label renders the event for explain/doc (#381).
-func (m *PTYMouse) Label() string {
-	verb := m.ButtonOrDefault()
-	if a := m.ActionOrDefault(); a != "click" || m.IsWheel() {
-		verb += "-" + a
-	}
-	if len(m.Mods) > 0 {
-		verb = strings.Join(m.Mods, "+") + "+" + verb
-	}
-	return fmt.Sprintf("%s at (%d,%d)", verb, m.Row, m.Col)
-}
-
-// MaxPTYSendTimes bounds the repeat count (#377). Navigation never needs
-// anywhere near it, and the cap keeps a typo'd `times: 100000000` from writing
-// gigabytes into the terminal instead of failing at the mistake.
-const MaxPTYSendTimes = 10000
-
-// The xterm bracketed-paste markers, i.e. what a terminal wraps pasted input in
-// once the application has turned DEC private mode 2004 on (#378). A terminal
-// emits them ONLY after that request, which is why sending a paste is gated on
-// having seen the program ask.
-const (
-	PasteStart = "\x1b[200~"
-	PasteEnd   = "\x1b[201~"
-)
-
 // PTYExpectScreen is a session-local rendered-screen wait: the matcher runs on
 // the live terminal screen during a pty session, not only after the program
 // exits. It reuses the StreamAssert surface (line/contains/matches/equals/json/
@@ -399,400 +112,4 @@ type PTYExpectScreen struct {
 	// this duration before the action passes, absorbing redraw churn without a
 	// blind sleep.
 	StableFor string `yaml:"stable_for,omitempty"`
-}
-
-// SendText is sugar for authoring the scalar form in Go literals (tests).
-func SendText(s string) *PTYSend { return &PTYSend{Text: &s} }
-
-// UnmarshalYAML decodes send as a scalar string or a {key: name} mapping,
-// rejecting unknown mapping keys (a custom unmarshaler bypasses the loader's
-// strict decode). It decodes from the AST node so every shape error carries
-// the offending value's [line:col] for the loader's excerpt formatter.
-func (p *PTYSend) UnmarshalYAML(node ast.Node) error {
-	fail := func(format string, args ...any) error {
-		return &yaml.SyntaxError{Message: fmt.Sprintf(format, args...), Token: node.GetToken()}
-	}
-	var one string
-	if err := yaml.NodeToValue(node, &one); err == nil {
-		p.Text = &one
-		return nil
-	}
-	var raw map[string]any
-	if err := yaml.NodeToValue(node, &raw); err != nil {
-		return fail("send must be a string or {key: <name>} (e.g. {key: enter})")
-	}
-	for k, v := range raw {
-		switch k {
-		case "key":
-			str, ok := v.(string)
-			if !ok {
-				return fail("send.key must be a string")
-			}
-			p.Key = strings.ToLower(strings.TrimSpace(str))
-		case "times":
-			// The range check lives here rather than in the loader because this
-			// is the only layer that can still tell an authored `times: 0` from
-			// an omitted one — the decoded field is a plain int, so by the time
-			// validation runs the two are the same value.
-			n, ok := asInt(v)
-			if !ok {
-				return fail("send.times must be an integer (e.g. {key: left, times: 16})")
-			}
-			if n < 1 {
-				return fail("send.times must be at least 1 (got %d); omit it for a single press", n)
-			}
-			if n > MaxPTYSendTimes {
-				return fail("send.times must not exceed %d (got %d)", MaxPTYSendTimes, n)
-			}
-			p.Times = n
-		case "paste":
-			str, ok := v.(string)
-			if !ok {
-				return fail("send.paste must be a string")
-			}
-			p.Paste = &str
-		case "mouse":
-			var mouse PTYMouse
-			raw, ok := v.(map[string]any)
-			if !ok {
-				return fail("send.mouse must be a mapping (e.g. {mouse: {row: 5, col: 12}})")
-			}
-			if err := decodeMouse(raw, &mouse, fail); err != nil {
-				return err
-			}
-			p.Mouse = &mouse
-		default:
-			return fail("send: unknown key %q (accepted: key, times, paste, mouse)", k)
-		}
-	}
-	if countSet(p.Key != "", p.Paste != nil, p.Mouse != nil) > 1 {
-		return fail("send: set exactly one of {key: <name>}, {paste: <text>}, or {mouse: {...}}")
-	}
-	if p.Mouse != nil {
-		if p.Times != 0 {
-			return fail("send.times repeats a named key; a mouse event is delivered once")
-		}
-		return nil
-	}
-	if p.Paste != nil {
-		if p.Times != 0 {
-			return fail("send.times repeats a named key; a paste is delivered once")
-		}
-		return nil
-	}
-	if p.Key == "" {
-		// This is also what rejects a lone `{times: N}`: repeating verbatim text
-		// is already expressible inline, so times only ever qualifies a key.
-		return fail("send: {key: <name>} requires a key name (e.g. enter, tab, ctrl-c), or use {paste: <text>}")
-	}
-	return nil
-}
-
-// countSet counts how many of the flags are true, for the one-of rules.
-func countSet(flags ...bool) int {
-	n := 0
-	for _, f := range flags {
-		if f {
-			n++
-		}
-	}
-	return n
-}
-
-// decodeMouse fills a PTYMouse from the decoded mapping, normalizing the
-// vocabularies to lower case. Membership is the loader's job; this only rejects
-// shapes (a row that is not a number, a mods list that is not a list of
-// strings), so every error still carries the offending node's [line:col].
-func decodeMouse(raw map[string]any, m *PTYMouse, fail func(string, ...any) error) error {
-	for k, v := range raw {
-		switch k {
-		case "row", "col":
-			n, ok := asInt(v)
-			if !ok {
-				return fail("send.mouse.%s must be an integer", k)
-			}
-			if k == "row" {
-				m.Row = n
-			} else {
-				m.Col = n
-			}
-		case "button", "action":
-			str, ok := v.(string)
-			if !ok {
-				return fail("send.mouse.%s must be a string", k)
-			}
-			norm := strings.ToLower(strings.TrimSpace(str))
-			if k == "button" {
-				m.Button = norm
-			} else {
-				m.Action = norm
-			}
-		case "mods":
-			list, ok := v.([]any)
-			if !ok {
-				return fail("send.mouse.mods must be a list (e.g. [ctrl])")
-			}
-			for _, item := range list {
-				str, ok := item.(string)
-				if !ok {
-					return fail("send.mouse.mods entries must be strings (accepted: %s)", PTYMouseModNames())
-				}
-				m.Mods = append(m.Mods, strings.ToLower(strings.TrimSpace(str)))
-			}
-		default:
-			return fail("send.mouse: unknown key %q (accepted: row, col, button, action, mods)", k)
-		}
-	}
-	return nil
-}
-
-// asInt accepts the several numeric shapes a YAML decoder may hand back for an
-// integer scalar, and rejects a float that is not whole (`times: 1.5` is a
-// mistake, not a truncation).
-func asInt(v any) (int, bool) {
-	switch n := v.(type) {
-	case int:
-		return n, true
-	case int64:
-		return int(n), true
-	case uint64:
-		if n > math.MaxInt32 {
-			return 0, false
-		}
-		return int(n), true
-	case float64:
-		if n != math.Trunc(n) {
-			return 0, false
-		}
-		return int(n), true
-	default:
-		return 0, false
-	}
-}
-
-// MarshalYAML emits the same shape UnmarshalYAML accepts — the scalar text form,
-// or the {key: <name>} mapping — so a loaded send round-trips back to a loadable
-// spec. Without it the default struct marshal writes a `text:` key the custom
-// unmarshaler rejects.
-func (p PTYSend) MarshalYAML() (any, error) {
-	if p.Text != nil {
-		return *p.Text, nil
-	}
-	if p.Paste != nil {
-		return map[string]string{"paste": *p.Paste}, nil
-	}
-	if p.Mouse != nil {
-		mouse := map[string]any{"row": p.Mouse.Row, "col": p.Mouse.Col}
-		if p.Mouse.Button != "" {
-			mouse["button"] = p.Mouse.Button
-		}
-		if p.Mouse.Action != "" {
-			mouse["action"] = p.Mouse.Action
-		}
-		if len(p.Mouse.Mods) > 0 {
-			mouse["mods"] = p.Mouse.Mods
-		}
-		return map[string]any{"mouse": mouse}, nil
-	}
-	if p.Times != 0 {
-		return map[string]any{"key": p.Key, "times": p.Times}, nil
-	}
-	return map[string]string{"key": p.Key}, nil
-}
-
-// ptyKeySequences maps each named key (#26) to the xterm byte sequence it
-// transmits. Documented bytes: enter=\r, tab=\t, esc=\x1b, space=" ",
-// backspace=\x7f (DEL, the modern erase), delete=\x1b[3~, insert=\x1b[2~,
-// arrows up/down/right/left=\x1b[A/B/C/D, home=\x1b[H, end=\x1b[F,
-// pageup=\x1b[5~, pagedown=\x1b[6~, f1-f4=\x1bOP..\x1bOS,
-// f5..f12=\x1b[15~,[17~..[21~,[23~,[24~, ctrl-a..ctrl-z=0x01..0x1a,
-// plus the punctuation aliases terminals conventionally expose for the
-// remaining C0 controls: ctrl-space/ctrl-@=NUL, ctrl-[=ESC, ctrl-\=FS,
-// ctrl-]=GS, ctrl-^=RS, ctrl-_=US. For modifier combos whose physical key does
-// not have a stable legacy C0 byte (e.g. Ctrl+-), use the terminal's CSI-u key
-// event instead so modern TUIs like Yazi see the intended modified key. ctrl-d
-// therefore stays the readable alias for the empty-send EOF rule.
-//
-// Three more families cover the chords real TUIs bind (#376), each the sequence
-// an xterm-class terminal sends for that physical key:
-//
-//   - shift-tab=\x1b[Z (CSI Z, "backtab"), the reverse-focus key every
-//     form-style TUI reads. `backtab` is an accepted alias for the same bytes.
-//   - alt-<letter>=\x1b<letter>, the ESC prefix a terminal transmits for a Meta
-//     chord (readline word operations, helix/emacs-style bindings), plus
-//     alt-enter=\x1b\r and alt-backspace=\x1b\x7f. Letters only: a Meta digit
-//     has no encoding stable enough to promise.
-//   - Modified arrows, in xterm's CSI 1 ; <modifier> <final> form —
-//     ctrl-<arrow>=\x1b[1;5A..D (word-wise cursor movement) and
-//     shift-<arrow>=\x1b[1;2A..D (selection movement).
-var ptyKeySequences = func() map[string]string {
-	m := map[string]string{
-		"enter":     "\r",
-		"tab":       "\t",
-		"esc":       "\x1b",
-		"space":     " ",
-		"backspace": "\x7f",
-		"delete":    "\x1b[3~",
-		"insert":    "\x1b[2~",
-		"shift-tab": "\x1b[Z",
-		"backtab":   "\x1b[Z",
-		"up":        "\x1b[A",
-		"down":      "\x1b[B",
-		"right":     "\x1b[C",
-		"left":      "\x1b[D",
-		"home":      "\x1b[H",
-		"end":       "\x1b[F",
-		"pageup":    "\x1b[5~",
-		"pagedown":  "\x1b[6~",
-		"f1":        "\x1bOP",
-		"f2":        "\x1bOQ",
-		"f3":        "\x1bOR",
-		"f4":        "\x1bOS",
-		"f5":        "\x1b[15~",
-		"f6":        "\x1b[17~",
-		"f7":        "\x1b[18~",
-		"f8":        "\x1b[19~",
-		"f9":        "\x1b[20~",
-		"f10":       "\x1b[21~",
-		"f11":       "\x1b[23~",
-		"f12":       "\x1b[24~",
-	}
-	for c := byte('a'); c <= 'z'; c++ {
-		m["ctrl-"+string(c)] = string([]byte{c - 'a' + 1})
-		m["alt-"+string(c)] = "\x1b" + string(c)
-	}
-	m["alt-enter"] = "\x1b\r"
-	m["alt-backspace"] = "\x1b\x7f"
-	// Modified arrows share one shape: CSI 1 ; <modifier> <final>, where the
-	// modifier is 1+bitmask (shift 1, alt 2, ctrl 4) — so ctrl is 5 and shift 2.
-	for i, final := range []string{"A", "B", "C", "D"} {
-		name := []string{"up", "down", "right", "left"}[i]
-		m["ctrl-"+name] = "\x1b[1;5" + final
-		m["shift-"+name] = "\x1b[1;2" + final
-	}
-	m["ctrl-space"] = "\x00"
-	m["ctrl-@"] = "\x00"
-	m["ctrl-["] = "\x1b"
-	m["ctrl-\\"] = "\x1c"
-	m["ctrl-]"] = "\x1d"
-	m["ctrl-^"] = "\x1e"
-	m["ctrl-_"] = "\x1f"
-	// xterm/kitty CSI-u for Ctrl+- (#286): raw 0x1f is Ctrl+_ and does not
-	// trigger TUIs that bind the physical hyphen key as a distinct modified key.
-	m["ctrl-hyphen"] = "\x1b[45;5u"
-	m["ctrl-minus"] = "\x1b[45;5u"
-	return m
-}()
-
-// ValidPTYKey reports whether name is in the named-key vocabulary (#26).
-func ValidPTYKey(name string) bool {
-	_, ok := ptyKeySequences[strings.ToLower(strings.TrimSpace(name))]
-	return ok
-}
-
-// ptyKeyBySequence reverse-maps an xterm byte sequence to its friendly key name
-// (#69), preferring the readable name over a ctrl-* alias when a byte is shared
-// (e.g. \r is both enter and ctrl-m — enter wins). Built once at init: the
-// ctrl-* aliases go in first, then the friendly names overwrite any collision.
-var ptyKeyBySequence = func() map[string]string {
-	m := make(map[string]string, len(ptyKeySequences))
-	m["\x00"] = "ctrl-space"
-	for c := byte('a'); c <= 'z'; c++ {
-		m[string([]byte{c - 'a' + 1})] = "ctrl-" + string(c)
-		// alt-<letter> is two bytes (ESC + the letter), so it never collides
-		// with a single-byte ctrl chord or with a CSI sequence. A user who
-		// pressed Esc and then typed the letter produces the same bytes, and
-		// replaying {key: alt-x} reproduces them exactly either way.
-		m["\x1b"+string(c)] = "alt-" + string(c)
-	}
-	m["\x1c"] = "ctrl-\\"
-	m["\x1d"] = "ctrl-]"
-	m["\x1e"] = "ctrl-^"
-	m["\x1f"] = "ctrl-_"
-	m["\x1b[45;5u"] = "ctrl-hyphen"
-	// backtab is deliberately absent: it and shift-tab share \x1b[Z, and a
-	// recording reads better naming the key a keyboard has.
-	for _, name := range []string{
-		"enter", "tab", "esc", "space", "backspace", "delete", "insert",
-		"shift-tab",
-		"up", "down", "right", "left", "home", "end",
-		"pageup", "pagedown",
-		"alt-enter", "alt-backspace",
-		"ctrl-up", "ctrl-down", "ctrl-right", "ctrl-left",
-		"shift-up", "shift-down", "shift-right", "shift-left",
-		"f1", "f2", "f3", "f4", "f5", "f6", "f7", "f8", "f9", "f10", "f11", "f12",
-	} {
-		m[ptyKeySequences[name]] = name
-	}
-	return m
-}()
-
-// PTYKeyForSequence returns the friendly named key whose xterm sequence exactly
-// equals seq (#69), so `atago record --pty` can render a lone control key as
-// {key: <name>} instead of an opaque escape. It reports false when no named key
-// matches the bytes exactly.
-func PTYKeyForSequence(seq string) (string, bool) {
-	name, ok := ptyKeyBySequence[seq]
-	return name, ok
-}
-
-// PTYKeyNames lists the vocabulary for error messages, compactly.
-func PTYKeyNames() string {
-	return "enter, tab, esc, space, backspace, delete, insert, up, down, left, right, home, end, pageup, pagedown, f1-f12, " +
-		"shift-tab (alias backtab), alt-a..alt-z, alt-enter, alt-backspace, " +
-		"ctrl-up/ctrl-down/ctrl-left/ctrl-right, shift-up/shift-down/shift-left/shift-right, " +
-		"ctrl-a..ctrl-z, ctrl-space/ctrl-@, ctrl-[, ctrl-\\, ctrl-], ctrl-^, ctrl-_, ctrl-hyphen/ctrl-minus"
-}
-
-// Bytes resolves the send payload to the bytes written to the terminal: the
-// named key's xterm sequence, the verbatim text, or 0x04 (VEOF, ^D) for the
-// historical empty-string EOF rule.
-func (p *PTYSend) Bytes() []byte {
-	if p.Mouse != nil {
-		return p.Mouse.Bytes()
-	}
-	if p.Paste != nil {
-		b := make([]byte, 0, len(PasteStart)+len(*p.Paste)+len(PasteEnd))
-		b = append(b, PasteStart...)
-		b = append(b, *p.Paste...)
-		return append(b, PasteEnd...)
-	}
-	if p.Key != "" {
-		seq := []byte(ptyKeySequences[p.Key])
-		if p.Times > 1 {
-			// One write, not p.Times writes: a held key reaches the program as a
-			// burst in a single read, and splitting it would let a redraw land
-			// between presses and change what the program sees.
-			return bytes.Repeat(seq, p.Times)
-		}
-		return seq
-	}
-	if p.Text != nil && *p.Text == "" {
-		return []byte{0x04}
-	}
-	if p.Text != nil {
-		return []byte(*p.Text)
-	}
-	return nil
-}
-
-// Label renders the send symbolically for explain/doc (#26): "press Enter"
-// for keys, a quoted excerpt for text, "EOF (^D)" for the empty string.
-func (p *PTYSend) Label() string {
-	switch {
-	case p.Mouse != nil:
-		return p.Mouse.Label()
-	case p.Paste != nil:
-		return fmt.Sprintf("paste %q", *p.Paste)
-	case p.Key != "" && p.Times > 1:
-		return fmt.Sprintf("press %s x%d", p.Key, p.Times)
-	case p.Key != "":
-		return "press " + p.Key
-	case p.Text != nil && *p.Text == "":
-		return "send EOF (^D)"
-	case p.Text != nil:
-		return fmt.Sprintf("type %q", *p.Text)
-	default:
-		return "send"
-	}
 }
